Fix entity ID type and add Nexus ID support in FCM notifications

+++ src/RustPlusApi.Fcm/Interfaces/IRustPlusFcm.cs
@@ -16,13 +16,13 @@ public interface IRustPlusFcm : IRustPlusFcmSocket
     event EventHandler<Notification<ServerEvent?>>? OnServerPairing;
 
     /// <summary>Raised when a smart switch pairing notification is received.</summary>
-    event EventHandler<Notification<int?>>? OnSmartSwitchPairing;
+    event EventHandler<Notification<ulong?>>? OnSmartSwitchPairing;
 
     /// <summary>Raised when a smart alarm pairing notification is received.</summary>
-    event EventHandler<Notification<int?>>? OnSmartAlarmPairing;
+    event EventHandler<Notification<ulong?>>? OnSmartAlarmPairing;
 
     /// <summary>Raised when a storage monitor pairing notification is received.</summary>
-    event EventHandler<Notification<int?>>? OnStorageMonitorPairing;
+    event EventHandler<Notification<ulong?>>? OnStorageMonitorPairing;
 
     /// <summary>Raised when a smart alarm is triggered.</summary>
     event EventHandler<AlarmEvent?>? OnAlarmTriggered;
